chore(spec2cdk): refactor code generation with LibraryBuilder abstraction
Introduces a new LibraryBuilder base class to abstract the code generation
process for CDK libraries. This refactoring separates concerns between the
general library building logic and aws-cdk-lib-specific implementation.

Key changes:
- Create LibraryBuilder base class with generic service submodule support
- Extract AwsCdkLibBuilder as a specialized implementation
- Introduce ServiceSubmodule abstraction for better modularity
- Move selective imports logic to service-submodule
- Update TypeConverter to support partial type definitions for mixins
- Adjust import paths and file patterns configuration
