+++ tools/@aws-cdk/spec2cdk/lib/cdk/relationship-decider.ts
@@ -3,6 +3,7 @@ import * as naming from '../naming';
 import { namespaceFromResource, referenceInterfaceName, referenceInterfaceAttributeName, referencePropertyName, typeAliasPrefixFromResource } from '../naming';
 import { getReferenceProps } from './reference-props';
 import { log } from '../util';
+import { SelectiveImport } from './service-submodule';
 
 // For now we want relationships to be applied only for these services
 export const RELATIONSHIP_SERVICES: string[] = [];
@@ -20,23 +21,6 @@ export interface Relationship {
   readonly propName: string;
   /** Human friendly name of the reference type for error generation (e.g. "iam.IRoleRef") */
   readonly typeDisplayName: string;
-}
-
-/**
- * Represents a selective import statement for cross-module type references.
- * Used to import specific types from other CDK modules when relationships
- * are between different modules.
- */
-export interface SelectiveImport {
-  /** The module name to import from */
-  readonly moduleName: string;
-  /** Array of types that need to be imported */
-  readonly types: {
-    /** The original type name in the source module */
-    originalType: string;
-    /** The aliased name to avoid naming conflicts */
-    aliasedType: string;
-  }[];
 }
 
 /**
