+++ tools/@aws-cdk/spec2cdk/lib/cdk/resource-class.ts
@@ -43,7 +43,8 @@ import {
 } from '../naming';
 import { isDefined, splitDocumentation, maybeDeprecated } from '../util';
 import { findArnProperty } from './reference-props';
-import { SelectiveImport, RelationshipDecider } from './relationship-decider';
+import { RelationshipDecider } from './relationship-decider';
+import { SelectiveImport } from './service-submodule';
 
 export interface ITypeHost {
   typeFromSpecType(type: PropertyType): Type;
