+++ tools/@aws-cdk/spec2cdk/lib/cli/cli.ts
@@ -7,6 +7,7 @@
 import * as path from 'node:path';
 import { parseArgs } from 'node:util';
 import { PositionalArg, showHelp } from './help';
+import { AwsCdkLibBuilder } from '../cdk/aws-cdk-lib';
 import { GenerateModuleMap, GenerateOptions, generate, generateAll } from '../generate';
 import { log } from '../util';
 
@@ -93,16 +94,18 @@ export async function main(argv: string[]) {
   }
 
   const outputPath = outputDir ?? path.join(__dirname, '..', 'services');
-  const generatorOptions: GenerateOptions = {
+
+  const generatorOptions: GenerateOptions<typeof AwsCdkLibBuilder> = {
     outputPath,
-    filePatterns: {
-      resources: options.pattern,
-      augmentations: options.augmentations,
-      cannedMetrics: options.metrics,
-    },
     clearOutput: options['clear-output'],
-    augmentationsSupport: options['augmentations-support'],
     debug: options.debug as boolean,
+    builderProps: {
+      filePatterns: {
+        resources: options.pattern,
+        augmentations: options.augmentations,
+        cannedMetrics: options.metrics,
+      },
+    },
   };
 
   if (options.service?.length) {
