+++ tools/@aws-cdk/spec2cdk/lib/generate.ts
@@ -1,11 +1,14 @@
-import * as path from 'path';
+import * as path from 'node:path';
 import { loadAwsServiceSpec } from '@aws-cdk/aws-service-spec';
 import { DatabaseBuilder } from '@aws-cdk/service-spec-importers';
-import { SpecDatabase } from '@aws-cdk/service-spec-types';
+import { Service, SpecDatabase } from '@aws-cdk/service-spec-types';
 import { TypeScriptRenderer } from '@cdklabs/typewriter';
 import * as fs from 'fs-extra';
-import { AstBuilder, AstBuilderProps, DEFAULT_FILE_PATTERNS } from './cdk/ast';
+import { AwsCdkLibBuilder } from './cdk/aws-cdk-lib';
+import { LibraryBuilder } from './cdk/library-builder';
 import { queryDb, log, TsFileWriter } from './util';
+
+export type BuilderProps<T> = T extends new (first: infer P, ...args: any[]) => any ? P : never;
 
 export interface GenerateServiceRequest {
   /**
@@ -41,7 +44,9 @@ export interface GenerateModuleOptions {
   readonly services: GenerateServiceRequest[];
 }
 
-export interface GenerateOptions extends Pick<AstBuilderProps, 'filePatterns' | 'inCdkLib'> {
+type LBC = new (...args: any[]) => LibraryBuilder;
+
+export interface GenerateOptions<Builder extends LBC>{
   /**
    * Base path for generated files
    *
@@ -58,17 +63,28 @@ export interface GenerateOptions extends Pick<AstBuilderProps, 'filePatterns' | 
   readonly clearOutput?: boolean;
 
   /**
-   * Generate L2 stub support files for augmentations (only for testing)
-   *
-   * @default false
-   */
-  readonly augmentationsSupport?: boolean;
-
-  /**
    * Output debug messages
    * @default false
    */
   readonly debug?: boolean;
+
+  /**
+   * Additional builder options.
+   */
+  readonly builderProps?: Partial<BuilderProps<Builder>>;
+
+  /**
+   * A function that returns an AstBuilder instance.
+   * @default - The default AstBuilder is constructed
+   */
+  readonly astBuilder?: Builder;
+
+  /**
+   * Provide an already loaded spec database.
+   *
+   * @default - load the patched spec db
+   */
+  readonly db?: SpecDatabase;
 }
 
 export interface GenerateModuleMap {
@@ -96,10 +112,10 @@ export interface GeneratedServiceInfo {
  * @param modules A map of arbitrary module names to GenerateModuleOptions. This allows for flexible generation of different configurations at a time.
  * @param options Configure the code generation
  */
-export async function generate(modules: GenerateModuleMap, options: GenerateOptions) {
+export async function generate<Builder extends LBC = typeof AwsCdkLibBuilder>(modules: GenerateModuleMap, options: GenerateOptions<Builder>) {
   enableDebug(options);
-  const db = await loadPatchedSpec();
-  return generator(db, modules, options);
+  const db = options.db ?? await loadPatchedSpec();
+  return generator<Builder>(db, modules, options);
 }
 
 /**
@@ -124,7 +140,7 @@ export async function loadPatchedSpec(): Promise<SpecDatabase> {
  * @param outputPath Base path for generated files. Use `options.filePatterns` to configure more complex scenarios.
  * @param options Additional configuration
  */
-export async function generateAll(options: GenerateOptions) {
+export async function generateAll<Builder extends LBC = typeof AwsCdkLibBuilder>(options: GenerateOptions<Builder>) {
   enableDebug(options);
   const db = await loadAwsServiceSpec();
   const services = queryDb.getAllServices(db);
@@ -140,16 +156,16 @@ export async function generateAll(options: GenerateOptions) {
   await generator(db, modules, options);
 }
 
-function enableDebug(options: GenerateOptions) {
+function enableDebug(options: { debug?: boolean }) {
   if (options.debug) {
     process.env.DEBUG = '1';
   }
 }
 
-async function generator(
+async function generator<Builder extends LBC = typeof AwsCdkLibBuilder>(
   db: SpecDatabase,
   modules: { [name: string]: GenerateModuleOptions },
-  options: GenerateOptions,
+  options: GenerateOptions<Builder>,
 ): Promise<GenerateOutput> {
   const timeLabel = '🐢  Completed in';
   log.time(timeLabel);
@@ -165,16 +181,14 @@ async function generator(
     fs.removeSync(outputPath);
   }
 
-  const ast = new AstBuilder({
+  const LibBuilder = options.astBuilder ?? AwsCdkLibBuilder;
+  const ast = new LibBuilder({
     db,
-    filePatterns: {
-      ...DEFAULT_FILE_PATTERNS,
-      ...noUndefined(options.filePatterns),
-    },
-    inCdkLib: options.inCdkLib,
-  });
+    ...options.builderProps,
+  } as any);
 
-  const moduleResources: Record<string, Record<string, string>> = {};
+  const servicesPerModule: Record<string, Service[]> = {};
+  const resourcesPerModule: Record<string, Record<string, string>> = {};
 
   // Go through the module map
   log.info('Generating %i modules...', Object.keys(modules).length);
@@ -190,8 +204,14 @@ async function generator(
         deprecated: req.deprecated,
       });
 
-      moduleResources[moduleName] = moduleResources[moduleName] ?? {};
-      Object.assign(moduleResources[moduleName], submod.resources);
+      servicesPerModule[moduleName] ??= [];
+      servicesPerModule[moduleName].push(s);
+
+      resourcesPerModule[moduleName] ??= {};
+      resourcesPerModule[moduleName] ??= {};
+      for (const [res, type] of submod.resources.entries()) {
+        resourcesPerModule[moduleName][res] = type.name;
+      }
     }
   }
 
@@ -200,10 +220,11 @@ async function generator(
   const writer = new TsFileWriter(outputPath, renderer);
   ast.writeAll(writer);
 
-  const allResources = Object.values(moduleResources).flat().reduce(mergeObjects, {});
+  const allResources = Object.values(resourcesPerModule).flat().reduce(mergeObjects, {});
   log.info('Summary:');
-  log.info('  Service files:  %i', Object.values(moduleOutputFiles).flat().length);
-  log.info('  Resources:      %i', Object.keys(allResources).length);
+  log.info('  Services:     %i', Object.values(servicesPerModule).flat().length);
+  log.info('  Files:        %i', Object.values(moduleOutputFiles).flat().length);
+  log.info('  Resources:    %i', Object.keys(allResources).length);
   log.timeEnd(timeLabel);
 
   return {
@@ -212,7 +233,7 @@ async function generator(
         moduleName,
         {
           outputFiles: Array.from(moduleOutputFiles[moduleName]).sort(),
-          resources: moduleResources[moduleName],
+          resources: resourcesPerModule[moduleName],
         },
       ]),
       )),
@@ -224,11 +245,4 @@ function mergeObjects<T>(all: T, res: T) {
     ...all,
     ...res,
   };
-}
-
-function noUndefined<A extends object>(x: A | undefined): A | undefined {
-  if (!x) {
-    return undefined;
-  }
-  return Object.fromEntries(Object.entries(x).filter(([, v]) => v !== undefined)) as any;
 }
