+++ tools/@aws-cdk/spec2cdk/lib/util/ts-file-writer.ts
@@ -1,6 +1,7 @@
 import * as path from 'node:path';
 import { Module, TypeScriptRenderer } from '@cdklabs/typewriter';
 import * as fs from 'fs-extra';
+import * as log from './log';
 
 export interface IWriter {
   write(module: Module, filePath: string): string;
@@ -15,6 +16,7 @@ export class TsFileWriter implements IWriter {
   ) {}
 
   public write(module: Module, filePath: string): string {
+    log.debug(module.name, filePath, 'render');
     const fullPath = path.join(this.rootDir, filePath);
     fs.outputFileSync(fullPath, this.renderer.render(module));
     this.outputFiles.push(fullPath);
