+++ tools/@aws-cdk/spec2cdk/test/history.test.ts
@@ -1,7 +1,7 @@
 import { loadAwsServiceSpec } from '@aws-cdk/aws-service-spec';
 import { SpecDatabase } from '@aws-cdk/service-spec-types';
 import { IScope } from '@cdklabs/typewriter';
-import { AstBuilder } from '../lib/cdk/ast';
+import { AwsCdkLibBuilder } from '../lib/cdk/aws-cdk-lib';
 
 let db: SpecDatabase;
 
@@ -13,7 +13,7 @@ beforeAll(async () => {
 // To ensure backwards compatibility we will render previous types
 test('Previous types are rendered', () => {
   const resource = db.lookup('resource', 'cloudFormationType', 'equals', 'AWS::CloudFormation::StackSet')[0];
-  const info = new AstBuilder({ db }).addResource(resource);
+  const info = new AwsCdkLibBuilder({ db }).addResource(resource);
 
   const modName = '@aws-cdk/aws-cloudformation/aws-cloudformation';
   const stackSet = info.resourcesMod.module.tryFindType(`${modName}.CfnStackSet`) as unknown as IScope;
