+++ tools/@aws-cdk/spec2cdk/test/services.test.ts
@@ -1,7 +1,7 @@
 import { loadAwsServiceSpec } from '@aws-cdk/aws-service-spec';
 import { SpecDatabase } from '@aws-cdk/service-spec-types';
 import { TypeScriptRenderer } from '@cdklabs/typewriter';
-import { AstBuilder } from '../lib/cdk/ast';
+import { AwsCdkLibBuilder } from '../lib/cdk/aws-cdk-lib';
 
 const renderer = new TypeScriptRenderer();
 let db: SpecDatabase;
@@ -13,7 +13,7 @@ beforeAll(async () => {
 test('can codegen service with arbitrary suffix', () => {
   const service = db.lookup('service', 'name', 'equals', 'aws-kinesisanalyticsv2').only();
 
-  const module = new AstBuilder({ db }).addService(service, { nameSuffix: 'V2' }).resourcesMod.module;
+  const module = new AwsCdkLibBuilder({ db }).addService(service, { nameSuffix: 'V2' }).resourcesMod.module;
 
   const rendered = renderer.render(module);
